requirements.txt
PyRSMQ==0.4.3
redis==3.5.3
